Move common_operation.go and common_polling.go to the new packages (#7894)
* Move common_operation.go to the tpgresource package

* Replace CommonOperationWaiter

* Replace function OperationWait

* Move common_operation_test.go to the tpgresource package

* common_operation_test.go stays in the google package

* Move common_polling.go to the transport package

* Add docs

* Add tpgresource/common_operation.go to the convertion list

* Remove duplicate import package
