+++ tfplan2cai/converters/google/resources/sqladmin_operation.go
@@ -118,7 +118,7 @@ func SqlAdminOperationWaitTime(config *transport_tpg.Config, res interface{}, pr
 	if err := w.SetOp(op); err != nil {
 		return err
 	}
-	return OperationWait(w, activity, timeout, config.PollInterval)
+	return tpgresource.OperationWait(w, activity, timeout, config.PollInterval)
 }
 
 // SqlAdminOperationError wraps sqladmin.OperationError and implements the
